requirements.txt
Django==1.6
South==0.8.4
argparse==1.2.1
distribute==0.6.24
django-tastypie==0.10.0
ipython==1.1.0
pep8==1.4.6
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
virtual-node==0.0.4
wsgiref==0.1.2
